requirements.txt
anthropic>=0.42.0
boto3>=1.35.0
PyYAML>=6.0
requests>=2.31.0
rich>=13.7.0

# Dev / test
pytest>=8.0.0
moto[dynamodb]>=5.0.0
